refactor: prepare issue tab view
